requirements.txt
flask
openai
python-docx
pandas
openpyxl>=3.0.0
requests
beautifulsoup4
playwright
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
Flask-Session>=0.5.0
python-dotenv>=1.0.0
PyYAML>=6.0
